enhanced_agent/requirements.txt
dspy-ai>=2.0.0
requests>=2.31.0
pathlib>=1.0.1
typing>=3.7.4.3
asyncio>=3.4.3 